+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * None
 
 ### Fixed
-* Getting `Backlink` properties of unmanaged Realm objects will throw an error: "Using backlinks is only possible for managed objects." ([#1293](https://github.com/realm/realm-dart/pull/1293)).
+* Getting `Backlink` properties of unmanaged Realm objects will throw an error: "Using backlinks is only possible for managed objects" ([#1293](https://github.com/realm/realm-dart/pull/1293)).
 
 ### Compatibility
 * Realm Studio: 13.0.0 or later.